chore: extracted shared scripting util & used it for type conversion
